feat(v3): make renovation page the homepage
